Move Image::BufferSparse to Image::Sparse::Buffer
Initially, the BufferSparse class was placed in the Image namespace, to reside alongside Buffer, BufferScratch and BufferPreload.
However when the atomic BufferScratch was added, and it required its own Value and Voxel classes, it was placed in its own 'atomic/' sub-directory within lib/image.
Given that similar classes are also present for sparse image handling, within the Image::Sparse namespace and in 'lib/image/sparse/', it made sense that the sparse image buffer should also live with these files and features.